Move setting back to IndexSettings
